Learned to handle sub-directories and append to existing projects

+++ WizardImplementationClass.cs
@@ -1,18 +1,18 @@
 ﻿// Based on https://docs.microsoft.com/en-us/visualstudio/extensibility/how-to-use-wizards-with-project-templates
 //      and https://stackoverflow.com/questions/3882764/issue-with-visual-studio-template-directory-creation
 
-using System;
-using System.IO;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
-using EnvDTE;
-using Microsoft.VisualStudio.TemplateWizard;
-using Microsoft.VisualStudio.Shell;
-using Microsoft.VisualStudio.Shell.Interop;
-using EnvDTE80;
-
 namespace VSCMakeWizards
 {
+    using System;
+    using EnvDTE;
+    using EnvDTE80;
+    using Microsoft.VisualStudio.Shell;
+    using Microsoft.VisualStudio.Shell.Interop;
+    using Microsoft.VisualStudio.TemplateWizard;
+    using System.Collections.Generic;
+    using System.IO;
+    using System.Text.RegularExpressions;
+
     public class WizardImplementation : IWizard
     {
         // This method is called before opening any item that   
@@ -37,9 +37,11 @@ namespace VSCMakeWizards
         {
         }
 
-        private void AddTemplate(string srcPath, 
-            string destPath, 
-            Dictionary<string, string> replacementsDictionary)
+        protected void AddTemplate(
+            string srcPath,
+            string destPath,
+            Dictionary<string, string> replacementsDictionary,
+            bool bAppend = false)
         {
             if (File.Exists(srcPath))
             {
@@ -49,49 +51,73 @@ namespace VSCMakeWizards
                 string input = File.ReadAllText(srcPath);
                 string output = re.Replace(input, match => replacementsDictionary[match.Groups[1].Value]);
 
-                File.WriteAllText(destPath, output);
+                // Don't overwrite
+                if (!File.Exists(destPath))
+                {
+                    File.WriteAllText(destPath, output);
+                }
+                else if (bAppend)
+                {
+                    File.AppendAllText(destPath, output);
+                }
             }
+        }
+
+        protected virtual void AddCMakeCode(string templatePath,
+            Dictionary<string, string> replacementsDictionary)
+        {
+        }
+
+        protected virtual void AddSourceFiles(string templatePath,
+            Dictionary<string, string> replacementsDictionary)
+        {
         }
 
         public void RunStarted(object automationObject,
             Dictionary<string, string> replacementsDictionary,
             WizardRunKind runKind, object[] customParams)
         {
-            var destinationDir = replacementsDictionary["$destinationdirectory$"];
             var desiredNamespace = replacementsDictionary["$safeprojectname$"];
             var templatePath = Path.GetDirectoryName((string)customParams[0]);
 
-            var dte = automationObject as DTE2;
-            var solution = dte.Solution as EnvDTE100.Solution4;
-
-            if (solution.IsOpen)
+            if (replacementsDictionary["$exclusiveproject$"] == Boolean.TrueString)
             {
-                solution.Close();
+                var dte = automationObject as DTE2;
+                var solution = dte.Solution as EnvDTE100.Solution4;
+
+                if (solution.IsOpen)
+                {
+                    solution.Close();
+                }
+            }
+
+            var solutionDir = replacementsDictionary["$solutiondirectory$"];
+
+            // If no solution name give, we take the project's safe name
+            if (!replacementsDictionary.ContainsKey("$specifiedsolutionname$"))
+            {
+                replacementsDictionary.Add("$specifiedsolutionname$", desiredNamespace);
+            }
+            else if (replacementsDictionary["$specifiedsolutionname$"] == string.Empty)
+            {
+                replacementsDictionary["$specifiedsolutionname$"] = desiredNamespace;
             }
 
             AddTemplate(Path.Combine(templatePath, "CMakeLists.txt"),
-                        Path.Combine(destinationDir, "CMakeLists.txt"),
+                        Path.Combine(solutionDir, "CMakeLists.txt"),
                         replacementsDictionary);
-            AddTemplate(Path.Combine(templatePath, "CMakeSettings.json"), 
-                        Path.Combine(destinationDir, "CMakeSettings.json"),
+            AddTemplate(Path.Combine(templatePath, "CMakeSettings.json"),
+                        Path.Combine(solutionDir, "CMakeSettings.json"),
                         replacementsDictionary);
-            // executable template
-            AddTemplate(Path.Combine(templatePath, "main.cpp"), 
-                        Path.Combine(destinationDir, desiredNamespace + ".cpp"),
-                        replacementsDictionary);
-            // library template
-            AddTemplate(Path.Combine(templatePath, "lib.cpp"),
-                        Path.Combine(destinationDir, desiredNamespace + ".cpp"),
-                        replacementsDictionary);
-            AddTemplate(Path.Combine(templatePath, "lib.h"),
-                        Path.Combine(destinationDir, desiredNamespace + ".h"),
-                        replacementsDictionary);
+
+            AddCMakeCode(templatePath, replacementsDictionary);
+            AddSourceFiles(templatePath, replacementsDictionary);
 
             var vsSolution = Package.GetGlobalService(typeof(SVsSolution)) as IVsSolution7;
 
             if (vsSolution != null)
             {
-                vsSolution.OpenFolder(destinationDir);
+                vsSolution.OpenFolder(solutionDir);
             }
 
             throw new WizardCancelledException();
@@ -102,6 +128,98 @@ namespace VSCMakeWizards
         public bool ShouldAddProjectItem(string filePath)
         {
             return false;
+        }
+    }
+
+    public class WizardImplementationExecutable : WizardImplementation
+    {
+        protected override void AddCMakeCode(string templatePath,
+            Dictionary<string, string> replacementsDictionary)
+        {
+            var solutionDir = replacementsDictionary["$solutiondirectory$"];
+            var destinationDir = replacementsDictionary["$destinationdirectory$"];
+
+            if (destinationDir == solutionDir)
+            {
+                AddTemplate(Path.Combine(templatePath, "AddExecutable.cmake"),
+                            Path.Combine(solutionDir, "CMakeLists.txt"),
+                            replacementsDictionary,
+                            true);
+            }
+            else
+            {
+                AddTemplate(Path.Combine(templatePath, "AddSubdirectory.cmake"),
+                            Path.Combine(solutionDir, "CMakeLists.txt"),
+                            replacementsDictionary,
+                            true);
+                AddTemplate(Path.Combine(templatePath, "AddExecutable.cmake"),
+                            Path.Combine(destinationDir, "CMakeLists.txt"),
+                            replacementsDictionary);
+            }
+        }
+
+        protected override void AddSourceFiles(string templatePath,
+            Dictionary<string, string> replacementsDictionary)
+        {
+            var destinationDir = replacementsDictionary["$destinationdirectory$"];
+            var destinationDirSrc = Path.Combine(destinationDir, "src");
+
+            Directory.CreateDirectory(destinationDirSrc);
+
+            var projectName = replacementsDictionary["$projectname$"];
+
+            AddTemplate(Path.Combine(templatePath, "main.cpp"),
+                        Path.Combine(destinationDirSrc, projectName + ".cpp"),
+                        replacementsDictionary);
+        }
+    }
+
+    public class WizardImplementationLibrary : WizardImplementation
+    {
+        protected override void AddCMakeCode(string templatePath,
+            Dictionary<string, string> replacementsDictionary)
+        {
+            var solutionDir = replacementsDictionary["$solutiondirectory$"];
+            var destinationDir = replacementsDictionary["$destinationdirectory$"];
+
+            if (destinationDir == solutionDir)
+            {
+                AddTemplate(Path.Combine(templatePath, "AddLibrary.cmake"),
+                            Path.Combine(solutionDir, "CMakeLists.txt"),
+                            replacementsDictionary,
+                            true);
+            }
+            else
+            {
+                AddTemplate(Path.Combine(templatePath, "AddSubdirectory.cmake"),
+                            Path.Combine(solutionDir, "CMakeLists.txt"),
+                            replacementsDictionary,
+                            true);
+                AddTemplate(Path.Combine(templatePath, "AddLibrary.cmake"),
+                            Path.Combine(destinationDir, "CMakeLists.txt"),
+                            replacementsDictionary);
+            }
+        }
+
+        protected override void AddSourceFiles(string templatePath,
+            Dictionary<string, string> replacementsDictionary)
+        {
+            var destinationDir = replacementsDictionary["$destinationdirectory$"];
+            var destinationDirSrc = Path.Combine(destinationDir, "src");
+            var destinationDirInc = Path.Combine(destinationDir, "inc");
+
+            Directory.CreateDirectory(destinationDirSrc);
+            Directory.CreateDirectory(destinationDirInc);
+
+            var projectName = replacementsDictionary["$projectname$"];
+
+            // dont't overwrite
+            AddTemplate(Path.Combine(templatePath, "lib.cpp"),
+                        Path.Combine(destinationDirSrc, projectName + ".cpp"),
+                        replacementsDictionary);
+            AddTemplate(Path.Combine(templatePath, "lib.h"),
+                        Path.Combine(destinationDirInc, projectName + ".h"),
+                        replacementsDictionary);
         }
     }
 }
